Cleaned up code and made some changes to the MUI
