+++ src/redux/sagas/login.saga.js
@@ -11,7 +11,8 @@ function* loginUser(action) {
       headers: { 'Content-Type': 'application/json' },
       withCredentials: true,
     };
-
+    console.log(config)
+    console.log(action.payload)
     // send the action.payload as the body
     // the config includes credentials which
     // allow the server session to recognize the user
