+++ src/templateComponents/App/App.css
@@ -55,7 +55,7 @@ Button
 }
 
 .btn:hover {
-  background-color: #008183;
+  background-color: #3D71B2;
   text-decoration: none;
 }
